requirements.txt
pytest==8.3.2
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pytest-django==4.8.0
django-extensions==3.2.3
drf_yasg==1.21.8